sync azure identity
